update teh product viewset view.

+++ products/views.py
@@ -6,6 +6,8 @@ from .serializers import ProductSerializer, ActivityLogSerializer, UserSerialize
 from django.utils.timezone import now
 from django.contrib.auth.models import User
 from rest_framework import generics
+from django.core.exceptions import PermissionDenied
+
 
 
 
@@ -16,18 +18,31 @@ class RegisterUserView(generics.CreateAPIView):
 
 
 class ProductViewSet(viewsets.ModelViewSet):
-    queryset = Product.objects.all()
+    queryset = Product.objects.all()  # 🔥 Add this line
     serializer_class = ProductSerializer
     permission_classes = [IsAuthenticated]
 
+    def get_queryset(self):
+        """Only return products created by the authenticated user."""
+        return Product.objects.filter(user=self.request.user)
+
     def perform_create(self, serializer):
+        """Assign the authenticated user as the product owner."""
         serializer.save(user=self.request.user)
 
     def perform_update(self, serializer):
+        """Ensure the user can only update their own products."""
+        instance = self.get_object()
+        if instance.user != self.request.user:
+            raise PermissionDenied("You do not have permission to update this product.")
         serializer.save()
 
     def perform_destroy(self, instance):
+        """Ensure the user can only delete their own products."""
+        if instance.user != self.request.user:
+            raise PermissionDenied("You do not have permission to delete this product.")
         instance.delete()
+
 
 
 class ActivityLogViewSet(viewsets.ReadOnlyModelViewSet):
